refactor: button to template syntax
